requirements.txt
# Core dependencies
anthropic>=0.39.0
openai>=1.0.0
python-dotenv>=1.0.0

# Google Workspace
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0

# Cloud functions
modal>=0.57.0

# Utilities
python-dateutil>=2.8.2
